requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
mysqlclient==2.2.7
psycopg==3.3.0
psycopg-binary==3.3.0
sqlparse==0.5.3
tzdata==2025.2
